feat:refactered user schema to take more inputs, refactored controllers and route responses pertaining to auth and users

+++ server/controller/logoutController.js
@@ -1,13 +1,17 @@
-const jwt = require('jsonwebtoken')
+const jwt = require("jsonwebtoken");
 const RefreshToken = require("../model/refreshToken");
 const handleLogout = async (req, res) => {
   try {
     const { refreshToken } = req.body;
+    console.log(refreshToken, "token")
     // Verify refresh token
     const decoded = jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET);
-
-    // Delete refresh token from database
-    await RefreshToken.findOneAndDelete({ token: refreshToken });
+    if (decoded) {
+      // Delete refresh token from database
+      await RefreshToken.findOneAndDelete({ token: refreshToken });
+    } else {
+      res.status(400).json({ error: "Invalid Refresh token" });
+    }
 
     res.json({ message: "Logout successful" });
   } catch (err) {
